+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # CHANGELOG
 
+## 1.0.0 - 2022-08-18
+
+The repository was archived.
+
+* Added `Josantonius\Validate\Validate::asDatetime()` method.
+* Added `Josantonius\Validate\Validate::hasMinLength()` method.
+* Added `Josantonius\Validate\Validate::hasSameLength()` method.
+* Added `Josantonius\Validate\Validate::exceedsMaxLength()` method.
+* Added `Josantonius\Validate\Validate::isGreaterThan()` method.
+* Added `Josantonius\Validate\Validate::isLessThan()` method.
+* Added `Josantonius\Validate\Validate::isEqualThan()` method.
+* Added `Josantonius\Validate\Validate::stringIsComposedBy()` method.
+* Added `Josantonius\Validate\Validate::stringIsComposedBy()` method.
+
+No tests were performed for the above methods.
+
 ## 1.0.0 - 2018-09-19
 
 * Added `Josantonius\Validate\Validate` class.
@@ -13,4 +29,3 @@
 * Added `Josantonius\Validate\Validate::asIp()` method.
 * Added `Josantonius\Validate\Validate::asUrl()` method.
 * Added `Josantonius\Validate\Validate::asEmail()` method.
-